backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.7.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
email-validator>=2.1.0.post1
greenlet>=3.1.1
bcrypt>=4.2.1